Fix lilypond-test
lilypond-test uses our texi2dvi, and texi2dvi uses etex.
As we don't have an own etex, the system's etex binary
is used.

Our texi2dvi needs a proper LD_LIBRARY_PATH, but etex
might fail (and it does so on openSuSE tumbleweed)
if LD_LIBRARY_PATH is not empty.

To solve this problem a script etex is created in
./target/tools/root/usr/bin. This script simply clears
LD_LIBRARY_PATH and then executes the system's
etex binary.